rename utils to lib folder
