+++ x-pack/plugins/search_playground/server/lib/assist.ts
