+++ x-pack/plugins/search_playground/server/lib/conversational_chain.ts
