+++ x-pack/plugins/search_playground/server/lib/elasticsearch_retriever.ts
