+++ x-pack/plugins/search_playground/server/lib/error_handler.ts
