+++ x-pack/plugins/search_playground/server/lib/fetch_indices.ts
