+++ x-pack/plugins/search_playground/server/lib/fetch_query_source_fields.ts
