+++ x-pack/plugins/search_playground/server/lib/get_chat_params.ts
