+++ x-pack/plugins/search_playground/server/lib/get_citations.ts
