+++ x-pack/plugins/search_playground/server/lib/render_template.ts
